Revert "vfio-pci: Allow to expose MSI-X table to userspace if interrupt remapping is enabled"
This reverts commit b9660f6b60b6e3446e1b5bd27ef0593907df12ec.

The original patch is not compatible with a recent change introduced
by upstream commit c535d34, since with that change in place we now
end up exiting msix_sparse_mmap_cap() before the
VFIO_REGION_INFO_CAP_SPARSE_MMAP flag is set for the region. This
results in us reverting back to not being able to mmap MMIO regions
for VFIO devices.

Signed-off-by: Michael Roth <mdroth@linux.vnet.ibm.com>