[IN REVIEW] arch/x86/amd/ibs: Fix re-arming IBS Fetch
Stephane Eranian “found that IBS FetchCtl was quickly getting stuck
with CurCnt not being reset to 0 during interrupt handling even though
the wrmsr to clear was executed."  He "then realized that if
ibsfetch.enable = 1, then a wrmsr to reset CurCnt has no effect.
In other words, to re-arm a period, IBS must be stopped.”

The PPR for AMD Family 17h Model 31h B0 55803 Rev 0.54 - Sep 12, 2019
states "The periodic fetch counter is set to IbsFetchCnt [...] when
IbsFetchEn is changed from 0 to 1."

Explicitly set IbsFetchEn to 0 and then to 1 when enabling IBS Fetch,
so the driver properly resets the internal counter to 0 and IBS
Fetch starts counting again.

It is not clear what versions of IBS hardware explicitly need
IbsFetchEn toggled and which historically may not have, so the
driver now always does it.

Reported-by: Stephane Eranian <stephane.eranian@google.com>
Signed-off-by: Kim Phillips <kim.phillips@amd.com>
Link: https://bugzilla.kernel.org/show_bug.cgi?id=206537
Cc: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Cc: Arnaldo Carvalho de Melo <acme@kernel.org>
Cc: Borislav Petkov <bp@alien8.de>
Cc: "H. Peter Anvin" <hpa@zytor.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Jiri Olsa <jolsa@redhat.com>
Cc: Mark Rutland <mark.rutland@arm.com>
Cc: Namhyung Kim <namhyung@kernel.org>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: linux-kernel@vger.kernel.org
Cc: x86@kernel.org
Cc: stable@vger.kernel.org