Early warning for BUG/WARN executing a UD2 instruction
The Cores team need to know whether a UD2 they see too late came
from one of the kernel's BUG or WARN assertions.

If warn_ud2 is set on the kernel command line, any BUG or WARN will
issue a warning and dump the stack on dmesg for serial capture
prior to continuing to execute the ud2 instruction.

At boot, a console message will always say whether this UD2 warning
feature is enabled.  It is off by default.

Signed-off-by: Kim Phillips <kim.phillips@amd.com>
Signed-off-by: Kalpana Shetty <kalpana.shetty@amd.com>