[INTERNAL] drivers/misc: Add HSMP Driver
The HSMP driver is an experimental kernel module for providing userspace
and kernel access to the Host System Management Port on AMD systems.
Initially this module is supported only on AMD family 17h and family 19h.

See amd_hsmp.h for descriptions of the kernel API.
See amd_hsmp.c for details about the SysFS interface.

Signed-off-by: Nathan Fontenot <nafonten@amd.com>