(INTERNAL) KVM: SVM: Add broadcast TLB invalidation for SEV and SEV-ES guests
In order to support INVLPGB/TLBSYNC in guests without any KVM
interception, Global ASIDs are required for the guests.

A guest is said to use Global ASID if the ASID is the same on all the
vCPUs for the guest.

Set bit 7 in VMCB offset 0x90 to enable usage of INVLPGB and TLBSYNC in
SEV guests.

Please refer to Appendix B Layout of VMCB in AMD64 Architecture
Programmer's Manual Volume 2.
https://www.amd.com/system/files/TechDocs/24593.pdf.

-Enable SEV guests to use INVLPGB/TLBSYNC -If host supports this feature
-Guest' CPUID has INVLPGB feature enabled -Guest kernel command line
should have tlbi=on/off.

Only SEV and higher guests currently use Global ASIDs.

Enable INVLPGB CPUID capability only for SEV and higher guests. Other
guests will see #UD when these instructions are executed.

Problem: A QEMU guest launched with -nosev and +invlpgb sees the INVLPGB
capability (CPU_80000008_EBX[3]) but when it uses INVLPGB it sees a #UD.

Solution :- If a guest CPUID has this capability but it does not use
global ASID, clear this capability when finalizing cpuids otherwise set
VMCB capability so such guests are able to use INVLPGB.

[ ms:
      - resolved merge conflicts
      - misc cleanups ]

Signed-off-by: Ramakrishna Saripalli <rk.saripalli@amd.com>
Co-developed-by: Manali Shukla <manali.shukla@amd.com>
Signed-off-by: Manali Shukla <manali.shukla@amd.com>