(INTERNAL) KVM: SVM: Add broadcast TLB invalidation for SEV and SEV-ES guests
In order to support INVLPGB/TLBSYNC in guests without any KVM
interception, Global ASIDs are required for the guests.

A guest is said to use Global ASID if the ASID is the same on all the
vCPUs for the guest.

Set bit 7 in VMCB offset 0x90 to enable usage of INVLPGB and TLBSYNC in
SEV guests.

Please refer to Appendix B Layout of VMCB in AMD64 Architecture
Programmer's Manual Volume 2.
https://www.amd.com/system/files/TechDocs/24593.pdf.

-Enable SEV guests to use INVLPGB/TLBSYNC -If host supports this feature
-Guest' CPUID has INVLPGB feature enabled -Guest kernel command line
should have tlbi=on/off.

Only SEV and higher guests currently use Global ASIDs.

Enable INVLPGB CPUID capability only for SEV and higher guests. Other
guests will see #UD when these instructions are executed.

Problem: A QEMU guest launched with -nosev and +invlpgb sees the INVLPGB
capability (CPU_80000008_EBX[3]) but when it uses INVLPGB it sees a #UD.

Solution :- If a guest CPUID has this capability but it does not use
global ASID, clear this capability when finalizing cpuids otherwise set
VMCB capability so such guests are able to use INVLPGB.

Performance (normalized) numbers on SEV and SEV-ES guests for in-house
benchmark:
---------------------------------------------------------------
| Genoa Host: 6.0.0+ Guest: 6.0.0+ 1P systems (SEV guest)     |
---------------------------------------------------------------
| No of | Elapsed time(s) | Elapsed time(s) | tlbi enabled    |
| cpus  | (tlbi=off)      | (tlbi=on)       | vs disabled (%) |
-----------------------------------------------------------------
|   2   |   1             |   0.167448868   |   83.25511321   |
|   4   |   1.769126007   |   0.167853131   |   90.51208731   |
|   8   |   3.215239172   |   0.168080667   |   94.77237437   |
|  16   |   6.225784254   |   0.168641775   |   97.29123644   |
|  32   |   11.94012046   |   0.173569146   |   98.54633673   |
|  64   |   6.513363325   |   0.182621989   |   97.19619528   |
-----------------------------------------------------------------

---------------------------------------------------------------
| Genoa Host: 6.0.0+ Guest: 6.0.0+ 1P systems (SEV-ES guest)  |
---------------------------------------------------------------
| No of | Elapsed time(s) | Elapsed time(s) | tlbi enabled    |
| cpus  | (tlbi=off)      | (tlbi=on)       | vs disabled (%) |
---------------------------------------------------------------
|   2   |   1             |   0.11898754    |   88.10124598   |
|   4   |   2.229660685   |   0.119546274   |   94.63836472   |
|   8   |   4.45644542    |   0.119854243   |   97.31054165   |
|  16   |   8.9109004     |   0.120624559   |   98.64632581   |
|  32   |   17.87093488   |   0.124191678   |   99.30506334   |
|  64   |   4.936211896   |   0.130785205   |   97.35049452   |
---------------------------------------------------------------

Signed-off-by: Ramakrishna Saripalli <rk.saripalli@amd.com>
Signed-off-by: Manali Shukla <manali.shukla@amd.com>