x86/sev-snp: define page state change VMGEXIT structure
An SNP-active guest will use the page state change VMGEXIT defined in
the GHCB specification section 4.1.6 to ask the hypervisor to make the
guest page private or shared in the RMP table. In addition to the
private/shared, the guest can also ask the hypervisor to split or
combine multiple 4K validated pages as a single 2M page or vice versa.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>