x86/sev-es: register GHCB memory when SEV-SNP is active
The SEV-SNP guest is required to perform GHCB GPA registration. This is
because the hypervisor may prefer that a guest use a consistent and/or
specific GPA for the GHCB associated with a vCPU. For more information,
see the GHCB specification section 2.5.2.

During the boot, init_ghcb() allocates a per-cpu GHCB page. On very first
VC exception, the execption handler switch to using the per-cpu GHCB page
allocated during the init_ghcb(). The GHCB page must be registered in
the vcpu context, so ensure that the per-cpu GHCB page is registered
before the first VMGEXIT from this vcpu.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>