x86: add the SEV-SNP initialization support
The memory integrity guarantees of a SEV-SNP are enforced through a new
structure called the Reverse Map Table (RMP). The RMP is a single data
structure shared across the system that contains one entry for every 4K
page of the DRAM that may be used by the VMs. The goal of RMP is to
track the owner for each page of the memory. Pages of memory can be
owned by the hypervisor, owned by a specific VM or owned by the AMD-SP.
See APM2 section 15.36.3 for more detail on RMP.

A RMP is used to enforce access control to memory, the table itself is
not directly writable by the software. New CPU instructions (RMPUPDATE,
PVALIDATE, RMPADJUST) can be used to manipulate the RMP entries.

Based on the platform configuration, the BIOS reserves the memory used
for the RMP table. The start and end address of the RMP table can be
queries by reading the RMP_BASE and RMP_END MSRs. If RMP_BASE and
RMP_END is not set then we disable the SEV-SNP feature.

Enable the SEV-SNP feature only after the RMP table initialization is
successful.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>