crypto ccp: handle the legacy SEV command when SNP is enabled
The behavior of the SEV-legacy commands are altered when the SNP firmware
is in the INIT state. When SNP is in INIT state, all the SEV-legacy
commands that causes the firmware to write to memory must be in the
firmware state before issuing the command..

See SEV-SNP spec section 5.3.7 for more detail.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>