KVM: SVM: make AVIC backing, VMSA and VMCB memory allocation SNP safe
When SEV-SNP is globally enabled on a system, the VMRUN instruction
performs additional security checks on AVIC backing, VMSA and VMCB page.
After, a successful VMRUN, these pages are marked "in-use" by the
hardware in the RMP entry, and any attempt to modify the RMP entry for
these pages will result in page-fault (RMP violation check).

While performing the RMP check, hardware will try to create a 2MB TLB
entry for the large page accessess. When it does this, it first reads
the RMP for the base of 2MB region and verify that all this memory is
safe. If AVIC backing, VMSA and VMCB memory happen to be the base of
2MB region, then RMP check will fail because of the "in-use" marking for
the base entry of this 2MB region.

e.g.

1. A VMCB was allocated on 2MB-aligned address.
2. The VMRUN instruction makred this RMP entry as "in-use".
3. Another process allocated some other page of memory that happened to be
   within the same 2MB region.
4. That process tried to write its page using physmap.

If the physmap entry in step #4 uses a large (1G/2M) page, then hardware
will attempt to create a 2M TLB entry. The hardware will find that the
"in-use" bit is set in the RMP entry (because it was a VMCB page) and will
cause a RMP violation check.

See APM2 section 15.36.12 for more information on VMRUN checks when
SEV-SNP is globally active.

A generic alloctor can return a page which are 2M aligned and will not be
safe to be used when SEV-SNP is globally enabled. Add
snp_safe_{alloc,free}_page() helper that can be used for allocating the
SNP safe memory. The helper allocated 2 pages, and splits them to order-1
allocation. It frees one page and keep one of the page whichever is not
2M aligned.

Co-developed-by: Marc Orr <marcorr@google.com>
Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>