KVM: SVM: add KVM_SEV_SNP_LAUNCH_START command
The KVM_SEV_SNP_LAUNCH_START begins the launch process. The command
initializes a cryptographic digest context used to construct the
measurement of the guest. If the guest is expected to be migrated,
the command also binds a migration agent (MA) to the guest.

For more information see the SEV-SNP spec section 4.5 and 8.11.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>