KVM: SVM: add KVM_SEV_SNP_LAUNCH_UPDATE command
The KVM_SEV_SNP_LAUNCH_UPDATE command can be used to insert data into the
guest's memory. The data is encrypted with the cryptographic context
created with the KVM_SEV_SNP_LAUNCH_START.

In addition to the inserting a data, it can insert a two special pages
into the guests memory: the secret page and the CPUID page.

For more information see the SEV-SNP spec section 4.5 and 8.12.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>