x86/mm: split the physmap when adding the page in RMP table
The integrity guarantee of SEV-SNP is enforced through RMP table. The RMP
is used in conjunctions with standard x86 page tables to enforce memory
restrictions and page access rights. The RMP is indexed by system
physical address, and is checked at the end of CPU and IOMMU table walk.
The RMP check is enforced as soon the SEV-SNP is enabled globally in the
system. Not every memory access requires an RMP check. In particular,
the read access from the hypervisor do not require RMP checks because
the data confidentiality is already protected via memory encryption
engine. When hardware encounters RMP check failure, it raise a page-fault
exception. The RMP bit in fault error code can be used to determine
if the fault was due to the RMP checks failure.

The write from the hypervisor goes through the RMP checks. When
hypervisor writes to pages, hardware checks to ensure that the assigned
bit in the RMP is zero (i.e page is shared). If the page table entry that
gives the sPA indicates that the target page size is large page, then all
RMP entries for the 4KB constituting pages of the target must have the
assigned bit 0. If one of entry does not have assigned bit 0 then hardware
will raise RMP violation. To resolve it, we must split the pages into 4K.

This poses a challenge in the Linux memory model. The Linux kernel
creates a direct mapping of all the physical memory -- referred as
physmap. The physmap may contain a valid mapping of guest owned pages.
During the page table walk, we may get into the situtation where one
of the page within the large page is owned by the guest (i.e assigned
bit is set in RMP). The write from the hypervisor will raise a RMP
violation. To workaround it, we call the set_memory_4k() to split the
physmap before adding the page in the RMP table. This ensure that the
pages added in the RMP table are used as 4K in the physmap.

The spliting of the physmap is a temporary solution until we work to
improve the kernel page fault handler to split the pages on demand.
One of the disadvtange of spliting is that eventually, we will end up
breaking the entire physmap unless we combine the split pages back to
large page. I am open for the suggestation on various approach we could
take to address this problem.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>