KVM: X86: define new RMP check related #NPF error bits
When SEV-SNP is enabld globally, the hardware places restrictions on all
memory accesses based on the RMP entry, whether the accessess are
performed by the hypervisor, a VM. When hardware encounters the RMP access
violation during the guest accessess, it will cause a #VMEXIT(NPF).
The new error bits can be used to determine the type of access.

See APM2 section 16.36.10 for more details.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>